TST: collect indexing tests by method (#40011)
