+++ pandas/tests/indexing/test_datetime.py
@@ -1,6 +1,3 @@
-import numpy as np
-import pytest
-
 import pandas as pd
 from pandas import (
     DataFrame,
@@ -54,28 +51,6 @@ class TestDatetimeIndex:
         result = df[df.a > df.a[3]]
         expected = df.iloc[4:]
         tm.assert_frame_equal(result, expected)
-
-    def test_setitem_with_expansion(self):
-        # indexing - setting an element
-        df = DataFrame(
-            data=pd.to_datetime(["2015-03-30 20:12:32", "2015-03-12 00:11:11"]),
-            columns=["time"],
-        )
-        df["new_col"] = ["new", "old"]
-        df.time = df.set_index("time").index.tz_localize("UTC")
-        v = df[df.new_col == "new"].set_index("time").index.tz_convert("US/Pacific")
-
-        # trying to set a single element on a part of a different timezone
-        # this converts to object
-        df2 = df.copy()
-        df2.loc[df2.new_col == "new", "time"] = v
-
-        expected = Series([v[0], df.loc[1, "time"]], name="time")
-        tm.assert_series_equal(df2.time, expected)
-
-        v = df.loc[df.new_col == "new", "time"] + pd.Timedelta("1s")
-        df.loc[df.new_col == "new", "time"] = v
-        tm.assert_series_equal(df.loc[df.new_col == "new", "time"], v)
 
     def test_consistency_with_tz_aware_scalar(self):
         # xef gh-12938
@@ -163,48 +138,6 @@ class TestDatetimeIndex:
         expected = Series([0, 5], index=index)
         tm.assert_series_equal(result, expected)
 
-    @pytest.mark.parametrize("to_period", [True, False])
-    def test_loc_getitem_listlike_of_datetimelike_keys(self, to_period):
-        # GH 11497
-
-        idx = date_range("2011-01-01", "2011-01-02", freq="D", name="idx")
-        if to_period:
-            idx = idx.to_period("D")
-        ser = Series([0.1, 0.2], index=idx, name="s")
-
-        keys = [Timestamp("2011-01-01"), Timestamp("2011-01-02")]
-        if to_period:
-            keys = [x.to_period("D") for x in keys]
-        result = ser.loc[keys]
-        exp = Series([0.1, 0.2], index=idx, name="s")
-        if not to_period:
-            exp.index = exp.index._with_freq(None)
-        tm.assert_series_equal(result, exp, check_index_type=True)
-
-        keys = [
-            Timestamp("2011-01-02"),
-            Timestamp("2011-01-02"),
-            Timestamp("2011-01-01"),
-        ]
-        if to_period:
-            keys = [x.to_period("D") for x in keys]
-        exp = Series(
-            [0.2, 0.2, 0.1], index=Index(keys, name="idx", dtype=idx.dtype), name="s"
-        )
-        result = ser.loc[keys]
-        tm.assert_series_equal(result, exp, check_index_type=True)
-
-        keys = [
-            Timestamp("2011-01-03"),
-            Timestamp("2011-01-02"),
-            Timestamp("2011-01-03"),
-        ]
-        if to_period:
-            keys = [x.to_period("D") for x in keys]
-
-        with pytest.raises(KeyError, match="with any missing labels"):
-            ser.loc[keys]
-
     def test_nanosecond_getitem_setitem_with_tz(self):
         # GH 11679
         data = ["2016-06-28 08:30:00.123456789"]
@@ -217,24 +150,6 @@ class TestDatetimeIndex:
         result = df.copy()
         result.loc[df.index[0], "a"] = -1
         expected = DataFrame(-1, index=index, columns=["a"])
-        tm.assert_frame_equal(result, expected)
-
-    def test_loc_setitem_with_expansion_and_existing_dst(self):
-        # GH 18308
-        start = Timestamp("2017-10-29 00:00:00+0200", tz="Europe/Madrid")
-        end = Timestamp("2017-10-29 03:00:00+0100", tz="Europe/Madrid")
-        ts = Timestamp("2016-10-10 03:00:00", tz="Europe/Madrid")
-        idx = pd.date_range(start, end, closed="left", freq="H")
-        assert ts not in idx  # i.e. result.loc setitem is with-expansion
-
-        result = DataFrame(index=idx, columns=["value"])
-        result.loc[ts, "value"] = 12
-        expected = DataFrame(
-            [np.nan] * len(idx) + [12],
-            index=idx.append(pd.DatetimeIndex([ts])),
-            columns=["value"],
-            dtype=object,
-        )
         tm.assert_frame_equal(result, expected)
 
     def test_getitem_millisecond_resolution(self, frame_or_series):
