+++ pandas/tests/indexing/test_floats.py
@@ -50,7 +50,7 @@ class TestFloatIndexers:
             tm.makePeriodIndex,
         ],
     )
-    def test_scalar_non_numeric(self, index_func, frame_or_series):
+    def test_scalar_non_numeric(self, index_func, frame_or_series, indexer_sl):
 
         # GH 4892
         # float_indexers should raise exceptions
@@ -61,10 +61,7 @@ class TestFloatIndexers:
 
         # getting
         with pytest.raises(KeyError, match="^3.0$"):
-            s[3.0]
-
-        with pytest.raises(KeyError, match="^3.0$"):
-            s.loc[3.0]
+            indexer_sl(s)[3.0]
 
         # contains
         assert 3.0 not in s
@@ -88,11 +85,7 @@ class TestFloatIndexers:
         else:
 
             s2 = s.copy()
-            s2.loc[3.0] = 10
-            assert s2.index.is_object()
-
-            s2 = s.copy()
-            s2[3.0] = 0
+            indexer_sl(s2)[3.0] = 10
             assert s2.index.is_object()
 
     @pytest.mark.parametrize(
@@ -114,7 +107,7 @@ class TestFloatIndexers:
         with pytest.raises(KeyError, match="^3.0$"):
             s[3.0]
 
-    def test_scalar_with_mixed(self):
+    def test_scalar_with_mixed(self, indexer_sl):
 
         s2 = Series([1, 2, 3], index=["a", "b", "c"])
         s3 = Series([1, 2, 3], index=["a", "b", 1.5])
@@ -122,36 +115,36 @@ class TestFloatIndexers:
         # lookup in a pure string index with an invalid indexer
 
         with pytest.raises(KeyError, match="^1.0$"):
-            s2[1.0]
+            indexer_sl(s2)[1.0]
 
         with pytest.raises(KeyError, match=r"^1\.0$"):
-            s2.loc[1.0]
+            indexer_sl(s2)[1.0]
 
-        result = s2.loc["b"]
+        result = indexer_sl(s2)["b"]
         expected = 2
         assert result == expected
 
         # mixed index so we have label
         # indexing
         with pytest.raises(KeyError, match="^1.0$"):
-            s3[1.0]
+            indexer_sl(s3)[1.0]
 
-        result = s3[1]
-        expected = 2
-        assert result == expected
+        if indexer_sl is not tm.loc:
+            # __getitem__ falls back to positional
+            result = s3[1]
+            expected = 2
+            assert result == expected
 
         with pytest.raises(KeyError, match=r"^1\.0$"):
-            s3.loc[1.0]
+            indexer_sl(s3)[1.0]
 
-        result = s3.loc[1.5]
+        result = indexer_sl(s3)[1.5]
         expected = 3
         assert result == expected
 
-    @pytest.mark.parametrize(
-        "idxr,getitem", [(lambda x: x.loc, False), (lambda x: x, True)]
-    )
     @pytest.mark.parametrize("index_func", [tm.makeIntIndex, tm.makeRangeIndex])
-    def test_scalar_integer(self, index_func, frame_or_series, idxr, getitem):
+    def test_scalar_integer(self, index_func, frame_or_series, indexer_sl):
+        getitem = indexer_sl is not tm.loc
 
         # test how scalar float indexers work on int indexes
 
@@ -161,7 +154,7 @@ class TestFloatIndexers:
 
         # coerce to equal int
 
-        result = idxr(obj)[3.0]
+        result = indexer_sl(obj)[3.0]
         self.check(result, obj, 3, getitem)
 
         if isinstance(obj, Series):
@@ -178,12 +171,12 @@ class TestFloatIndexers:
                 expected = Series(100.0, index=range(len(obj)), name=3)
 
         s2 = obj.copy()
-        idxr(s2)[3.0] = 100
+        indexer_sl(s2)[3.0] = 100
 
-        result = idxr(s2)[3.0]
+        result = indexer_sl(s2)[3.0]
         compare(result, expected)
 
-        result = idxr(s2)[3]
+        result = indexer_sl(s2)[3]
         compare(result, expected)
 
     @pytest.mark.parametrize("index_func", [tm.makeIntIndex, tm.makeRangeIndex])
@@ -204,7 +197,8 @@ class TestFloatIndexers:
 
         # assert all operations except for iloc are ok
         indexer = index[3]
-        for idxr, getitem in [(lambda x: x.loc, False), (lambda x: x, True)]:
+        for idxr in [tm.loc, tm.setitem]:
+            getitem = idxr is not tm.loc
 
             # getting
             result = idxr(s)[indexer]
@@ -242,7 +236,7 @@ class TestFloatIndexers:
         ],
     )
     @pytest.mark.parametrize("idx", [slice(3.0, 4), slice(3, 4.0), slice(3.0, 4.0)])
-    def test_slice_non_numeric(self, index_func, idx, frame_or_series):
+    def test_slice_non_numeric(self, index_func, idx, frame_or_series, indexer_sli):
 
         # GH 4892
         # float_indexers should raise exceptions
@@ -252,38 +246,28 @@ class TestFloatIndexers:
         s = gen_obj(frame_or_series, index)
 
         # getitem
-        msg = (
-            "cannot do positional indexing "
-            fr"on {type(index).__name__} with these indexers \[(3|4)\.0\] of "
-            "type float"
-        )
+        if indexer_sli is tm.iloc:
+            msg = (
+                "cannot do positional indexing "
+                fr"on {type(index).__name__} with these indexers \[(3|4)\.0\] of "
+                "type float"
+            )
+        else:
+            msg = (
+                "cannot do slice indexing "
+                fr"on {type(index).__name__} with these indexers "
+                r"\[(3|4)(\.0)?\] "
+                r"of type (float|int)"
+            )
         with pytest.raises(TypeError, match=msg):
-            s.iloc[idx]
-
-        msg = (
-            "cannot do (slice|positional) indexing "
-            fr"on {type(index).__name__} with these indexers "
-            r"\[(3|4)(\.0)?\] "
-            r"of type (float|int)"
-        )
-        for idxr in [lambda x: x.loc, lambda x: x.iloc, lambda x: x]:
-            with pytest.raises(TypeError, match=msg):
-                idxr(s)[idx]
+            indexer_sli(s)[idx]
 
         # setitem
-        msg = "slice indices must be integers or None or have an __index__ method"
+        if indexer_sli is tm.iloc:
+            # otherwise we keep the same message as above
+            msg = "slice indices must be integers or None or have an __index__ method"
         with pytest.raises(TypeError, match=msg):
-            s.iloc[idx] = 0
-
-        msg = (
-            "cannot do (slice|positional) indexing "
-            fr"on {type(index).__name__} with these indexers "
-            r"\[(3|4)(\.0)?\] "
-            r"of type (float|int)"
-        )
-        for idxr in [lambda x: x.loc, lambda x: x]:
-            with pytest.raises(TypeError, match=msg):
-                idxr(s)[idx] = 0
+            indexer_sli(s)[idx] = 0
 
     def test_slice_integer(self):
 
@@ -469,25 +453,24 @@ class TestFloatIndexers:
             s[idx]
 
     @pytest.mark.parametrize("idx", [slice(3.0, 4), slice(3, 4.0), slice(3.0, 4.0)])
-    def test_slice_float(self, idx, frame_or_series):
+    def test_slice_float(self, idx, frame_or_series, indexer_sl):
 
         # same as above, but for floats
         index = Index(np.arange(5.0)) + 0.1
         s = gen_obj(frame_or_series, index)
 
         expected = s.iloc[3:4]
-        for idxr in [lambda x: x.loc, lambda x: x]:
 
-            # getitem
-            result = idxr(s)[idx]
-            assert isinstance(result, type(s))
-            tm.assert_equal(result, expected)
+        # getitem
+        result = indexer_sl(s)[idx]
+        assert isinstance(result, type(s))
+        tm.assert_equal(result, expected)
 
-            # setitem
-            s2 = s.copy()
-            idxr(s2)[idx] = 0
-            result = idxr(s2)[idx].values.ravel()
-            assert (result == 0).all()
+        # setitem
+        s2 = s.copy()
+        indexer_sl(s2)[idx] = 0
+        result = indexer_sl(s2)[idx].values.ravel()
+        assert (result == 0).all()
 
     def test_floating_index_doc_example(self):
 
@@ -563,19 +546,6 @@ class TestFloatIndexers:
 
         result = indexer_sl(s)[[2.5]]
         tm.assert_series_equal(result, Series([1], index=[2.5]))
-
-    def test_floating_tuples(self):
-        # see gh-13509
-        s = Series([(1, 1), (2, 2), (3, 3)], index=[0.0, 0.1, 0.2], name="foo")
-
-        result = s[0.0]
-        assert result == (1, 1)
-
-        expected = Series([(1, 1), (2, 2)], index=[0.0, 0.0], name="foo")
-        s = Series([(1, 1), (2, 2), (3, 3)], index=[0.0, 0.0, 0.2], name="foo")
-
-        result = s[0.0]
-        tm.assert_series_equal(result, expected)
 
     def test_float64index_slicing_bug(self):
         # GH 5557, related to slicing a float index
