+++ pandas/tests/indexing/test_indexing.py
@@ -101,12 +101,12 @@ class TestFancy:
         idxr = indexer_sli(obj)
         nd3 = np.random.randint(5, size=(2, 2, 2))
 
-        if indexer_sli.__name__ == "iloc":
+        if indexer_sli is tm.iloc:
             err = ValueError
             msg = f"Cannot set values with ndim > {obj.ndim}"
         elif (
             isinstance(index, pd.IntervalIndex)
-            and indexer_sli.__name__ == "setitem"
+            and indexer_sli is tm.setitem
             and obj.ndim == 1
         ):
             err = AttributeError
@@ -136,17 +136,6 @@ class TestFancy:
 
         result = df.index
         expected = pd.Float64Index([1, 2, np.inf])
-        tm.assert_index_equal(result, expected)
-
-    def test_loc_setitem_with_expasnion_inf_upcast_empty(self):
-        # Test with np.inf in columns
-        df = DataFrame()
-        df.loc[0, 0] = 1
-        df.loc[1, 1] = 2
-        df.loc[0, np.inf] = 3
-
-        result = df.columns
-        expected = pd.Float64Index([0, 1, np.inf])
         tm.assert_index_equal(result, expected)
 
     def test_setitem_dtype_upcast(self):
@@ -308,12 +297,11 @@ class TestFancy:
         result = df.loc[[1, 2], ["a", "b"]]
         tm.assert_frame_equal(result, expected)
 
-    @pytest.mark.parametrize("case", [tm.getitem, tm.loc])
-    def test_duplicate_int_indexing(self, case):
+    def test_duplicate_int_indexing(self, indexer_sl):
         # GH 17347
         s = Series(range(3), index=[1, 1, 3])
         expected = s[1]
-        result = case(s)[[1]]
+        result = indexer_sl(s)[[1]]
         tm.assert_series_equal(result, expected)
 
     def test_indexing_mixed_frame_bug(self):
@@ -496,40 +484,6 @@ class TestFancy:
 
         result = DataFrame(index=[0, 1], columns=[0])
         result.iloc[1, 0] = [1, 2]
-
-        tm.assert_frame_equal(result, df)
-
-    def test_iloc_setitem_custom_object(self):
-        # iloc with an object
-        class TO:
-            def __init__(self, value):
-                self.value = value
-
-            def __str__(self) -> str:
-                return f"[{self.value}]"
-
-            __repr__ = __str__
-
-            def __eq__(self, other) -> bool:
-                return self.value == other.value
-
-            def view(self):
-                return self
-
-        df = DataFrame(index=[0, 1], columns=[0])
-        df.iloc[1, 0] = TO(1)
-        df.iloc[1, 0] = TO(2)
-
-        result = DataFrame(index=[0, 1], columns=[0])
-        result.iloc[1, 0] = TO(2)
-
-        tm.assert_frame_equal(result, df)
-
-        # remains object dtype even after setting it back
-        df = DataFrame(index=[0, 1], columns=[0])
-        df.iloc[1, 0] = TO(1)
-        df.iloc[1, 0] = np.nan
-        result = DataFrame(index=[0, 1], columns=[0])
 
         tm.assert_frame_equal(result, df)
 
@@ -748,7 +702,7 @@ class TestMisc:
         with pytest.raises(ValueError, match="slice step cannot be zero"):
             indexer_sl(ser)[::0]
 
-    def test_indexing_assignment_dict_already_exists(self):
+    def test_loc_setitem_indexing_assignment_dict_already_exists(self):
         index = Index([-5, 0, 5], name="z")
         df = DataFrame({"x": [1, 2, 6], "y": [2, 2, 8]}, index=index)
         expected = df.copy()
@@ -763,7 +717,7 @@ class TestMisc:
         expected = DataFrame({"x": [1, 2, 9], "y": [2.0, 2.0, 99.0]}, index=index)
         tm.assert_frame_equal(df, expected)
 
-    def test_indexing_dtypes_on_empty(self):
+    def test_iloc_getitem_indexing_dtypes_on_empty(self):
         # Check that .iloc returns correct dtypes GH9983
         df = DataFrame({"a": [1, 2, 3], "b": ["b", "b2", "b3"]})
         df2 = df.iloc[[], :]
@@ -772,7 +726,7 @@ class TestMisc:
         tm.assert_series_equal(df2.loc[:, "a"], df2.iloc[:, 0])
 
     @pytest.mark.parametrize("size", [5, 999999, 1000000])
-    def test_range_in_series_indexing(self, size):
+    def test_loc_range_in_series_indexing(self, size):
         # range can cause an indexing error
         # GH 11652
         s = Series(index=range(size), dtype=np.float64)
