+++ pandas/tests/series/indexing/test_indexing.py
@@ -82,16 +82,6 @@ def test_getitem_setitem_ellipsis():
     assert (result == 5).all()
 
 
-def test_setitem_with_expansion_type_promotion():
-    # GH12599
-    s = Series(dtype=object)
-    s["a"] = Timestamp("2016-01-01")
-    s["b"] = 3.0
-    s["c"] = "foo"
-    expected = Series([Timestamp("2016-01-01"), 3.0, "foo"], index=["a", "b", "c"])
-    tm.assert_series_equal(s, expected)
-
-
 @pytest.mark.parametrize(
     "result_1, duplicate_item, expected_1",
     [
@@ -193,38 +183,10 @@ def test_setitem_slicestep():
     assert (series[::2] == 0).all()
 
 
-def test_setitem_not_contained(string_series):
-    # set item that's not contained
-    ser = string_series.copy()
-    ser["foobar"] = 1
-
-    app = Series([1], index=["foobar"], name="series")
-    expected = string_series.append(app)
-    tm.assert_series_equal(ser, expected)
-
-
 def test_setslice(datetime_series):
     sl = datetime_series[5:20]
     assert len(sl) == len(sl.index)
     assert sl.index.is_unique is True
-
-
-def test_loc_setitem_2d_to_1d_raises():
-    x = np.random.randn(2, 2)
-    y = Series(range(2))
-
-    msg = "|".join(
-        [
-            r"shape mismatch: value array of shape \(2,2\)",
-            r"cannot reshape array of size 4 into shape \(2,\)",
-        ]
-    )
-    with pytest.raises(ValueError, match=msg):
-        y.loc[range(2)] = x
-
-    msg = r"could not broadcast input array from shape \(2,2\) into shape \(2,?\)"
-    with pytest.raises(ValueError, match=msg):
-        y.loc[:] = x
 
 
 # FutureWarning from NumPy about [slice(None, 5).
@@ -252,84 +214,7 @@ def test_basic_getitem_setitem_corner(datetime_series):
         datetime_series[[5, slice(None, None)]] = 2
 
 
-@pytest.mark.parametrize("tz", ["US/Eastern", "UTC", "Asia/Tokyo"])
-def test_setitem_with_tz(tz, indexer_sli):
-    orig = Series(pd.date_range("2016-01-01", freq="H", periods=3, tz=tz))
-    assert orig.dtype == f"datetime64[ns, {tz}]"
-
-    exp = Series(
-        [
-            Timestamp("2016-01-01 00:00", tz=tz),
-            Timestamp("2011-01-01 00:00", tz=tz),
-            Timestamp("2016-01-01 02:00", tz=tz),
-        ]
-    )
-
-    # scalar
-    ser = orig.copy()
-    indexer_sli(ser)[1] = Timestamp("2011-01-01", tz=tz)
-    tm.assert_series_equal(ser, exp)
-
-    # vector
-    vals = Series(
-        [Timestamp("2011-01-01", tz=tz), Timestamp("2012-01-01", tz=tz)],
-        index=[1, 2],
-    )
-    assert vals.dtype == f"datetime64[ns, {tz}]"
-
-    exp = Series(
-        [
-            Timestamp("2016-01-01 00:00", tz=tz),
-            Timestamp("2011-01-01 00:00", tz=tz),
-            Timestamp("2012-01-01 00:00", tz=tz),
-        ]
-    )
-
-    ser = orig.copy()
-    indexer_sli(ser)[[1, 2]] = vals
-    tm.assert_series_equal(ser, exp)
-
-
-def test_setitem_with_tz_dst(indexer_sli):
-    # GH XXX TODO: fill in GH ref
-    tz = "US/Eastern"
-    orig = Series(pd.date_range("2016-11-06", freq="H", periods=3, tz=tz))
-    assert orig.dtype == f"datetime64[ns, {tz}]"
-
-    exp = Series(
-        [
-            Timestamp("2016-11-06 00:00-04:00", tz=tz),
-            Timestamp("2011-01-01 00:00-05:00", tz=tz),
-            Timestamp("2016-11-06 01:00-05:00", tz=tz),
-        ]
-    )
-
-    # scalar
-    ser = orig.copy()
-    indexer_sli(ser)[1] = Timestamp("2011-01-01", tz=tz)
-    tm.assert_series_equal(ser, exp)
-
-    # vector
-    vals = Series(
-        [Timestamp("2011-01-01", tz=tz), Timestamp("2012-01-01", tz=tz)],
-        index=[1, 2],
-    )
-    assert vals.dtype == f"datetime64[ns, {tz}]"
-
-    exp = Series(
-        [
-            Timestamp("2016-11-06 00:00", tz=tz),
-            Timestamp("2011-01-01 00:00", tz=tz),
-            Timestamp("2012-01-01 00:00", tz=tz),
-        ]
-    )
-
-    ser = orig.copy()
-    indexer_sli(ser)[[1, 2]] = vals
-    tm.assert_series_equal(ser, exp)
-
-
-def test_categorical_assigning_ops():
+def test_setitem_categorical_assigning_ops():
     orig = Series(Categorical(["b", "b"], categories=["a", "b"]))
     s = orig.copy()
     s[:] = "a"
@@ -387,11 +272,6 @@ def test_slice(string_series, object_series):
     assert (string_series[10:20] == 0).all()
 
 
-def test_slice_can_reorder_not_uniquely_indexed():
-    s = Series(1, index=["a", "a", "b", "b", "c"])
-    s[::-1]  # it works!
-
-
 def test_loc_setitem(string_series):
     inds = string_series.index[[3, 4, 7]]
 
@@ -433,15 +313,6 @@ def test_timedelta_assignment():
     tm.assert_series_equal(s, expected)
 
 
-def test_setitem_td64_non_nano():
-    # GH 14155
-    ser = Series(10 * [np.timedelta64(10, "m")])
-    ser.loc[[1, 2, 3]] = np.timedelta64(20, "m")
-    expected = Series(10 * [np.timedelta64(10, "m")])
-    expected.loc[[1, 2, 3]] = Timedelta(np.timedelta64(20, "m"))
-    tm.assert_series_equal(ser, expected)
-
-
 def test_underlying_data_conversion():
     # GH 4080
     df = DataFrame({c: [1, 2, 3] for c in ["a", "b", "c"]})
@@ -457,33 +328,6 @@ def test_underlying_data_conversion():
     )
     return_value = expected.set_index(["a", "b", "c"], inplace=True)
     assert return_value is None
-    tm.assert_frame_equal(df, expected)
-
-
-def test_chained_assignment():
-    # GH 3970
-    with pd.option_context("chained_assignment", None):
-        df = DataFrame({"aa": range(5), "bb": [2.2] * 5})
-        df["cc"] = 0.0
-
-        ck = [True] * len(df)
-
-        df["bb"].iloc[0] = 0.13
-
-        # TODO: unused
-        df_tmp = df.iloc[ck]  # noqa
-
-        df["bb"].iloc[0] = 0.15
-        assert df["bb"].iloc[0] == 0.15
-
-
-def test_setitem_with_expansion_dtype():
-    # GH 3217
-    df = DataFrame({"a": [1, 3], "b": [np.nan, 2]})
-    df["c"] = np.nan
-    df["c"].update(Series(["foo"], index=[0]))
-
-    expected = DataFrame({"a": [1, 3], "b": [np.nan, 2], "c": ["foo", np.nan]})
     tm.assert_frame_equal(df, expected)
 
 
@@ -535,44 +379,21 @@ def test_setitem_mask_promote():
     tm.assert_series_equal(ser, expected)
 
 
-def test_multilevel_preserve_name():
+def test_multilevel_preserve_name(indexer_sl):
     index = MultiIndex(
         levels=[["foo", "bar", "baz", "qux"], ["one", "two", "three"]],
         codes=[[0, 0, 0, 1, 1, 2, 2, 3, 3, 3], [0, 1, 2, 0, 1, 1, 2, 0, 1, 2]],
         names=["first", "second"],
     )
-    s = Series(np.random.randn(len(index)), index=index, name="sth")
+    ser = Series(np.random.randn(len(index)), index=index, name="sth")
 
-    result = s["foo"]
-    result2 = s.loc["foo"]
-    assert result.name == s.name
-    assert result2.name == s.name
+    result = indexer_sl(ser)["foo"]
+    assert result.name == ser.name
 
 
 """
 miscellaneous methods
 """
-
-
-def test_uint_drop(any_int_dtype):
-    # see GH18311
-    # assigning series.loc[0] = 4 changed series.dtype to int
-    series = Series([1, 2, 3], dtype=any_int_dtype)
-    series.loc[0] = 4
-    expected = Series([4, 2, 3], dtype=any_int_dtype)
-    tm.assert_series_equal(series, expected)
-
-
-def test_getitem_unrecognized_scalar():
-    # GH#32684 a scalar key that is not recognized by lib.is_scalar
-
-    # a series that might be produced via `frame.dtypes`
-    ser = Series([1, 2], index=[np.dtype("O"), np.dtype("i8")])
-
-    key = ser.index[1]
-
-    result = ser[key]
-    assert result == 2
 
 
 def test_slice_with_zero_step_raises(index, frame_or_series, indexer_sli):
